+++ mutatest/_devtools.py
@@ -47,6 +47,12 @@ class NodeLister(ast.NodeVisitor):  # pragma: no cover
         print(ast.dump(node))
         self.generic_visit(node)
 
+    def visit_If(self, node: ast.If) -> None:
+        # have is, is not already, not clear If directly is a valuable mutation
+        print(f"If: {node}")
+        print(ast.dump(node))
+        self.generic_visit(node)
+
     def visit_Index(self, node: ast.Index) -> None:
         # actual slicing
         print(f"Index: {node}")
@@ -72,12 +78,6 @@ class NodeLister(ast.NodeVisitor):  # pragma: no cover
 
     # IN DEVELOPMENT
     ################################################################################################
-
-    def visit_If(self, node: ast.If) -> None:
-        # have is, is not already, not clear If directly is a valuable mutation
-        # print(f"If: {node}")
-        print(ast.dump(node))
-        self.generic_visit(node)
 
     # FUTURE DEVELOPMENT
     ################################################################################################